requirements.txt
opencv-python~=4.5.2.52
numpy~=1.19.5
Keras~=2.4.3
tensorflow~=2.5.0
h5py~=3.1.0
PySimpleGUI~=4.43.0
Pillow~=8.2.0